pyproject.toml
[tool.poetry]
name = "grpc-internal-communication"
version = "0.1.0"
description = ""
authors = ["Oliver Bielik <oliverbielik@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.73.1"
grpcio-tools = "^1.73.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
